Revamped the underlying models a lot
Used Backbone so that we get real classes. Ports are much better defined
and validate and check for cycles by default. They automatically
propagate their values through the system, so all the elements are
always up-to-date. Also made all the circuit elements Backbone classes
and shared a lot of code and stuff between them. Also importantly, made
the container component one large svg that takes up the whole screen, so
the drag handler can be placed on the entire svg and works much better.